Improve support functionality
