+++ support/validate_helper.js
@@ -1,6 +1,5 @@
-const liskCommons = require('lisk-commons');
 const chai = require('chai');
-const { TO_LISK, TO_BEDDOWS, from, sortBy, flattern } = require('../utils');
+const { TO_LISK, TO_BEDDOWS, from } = require('../utils');
 const LiskUtil = require('./lisk_util');
 
 /* eslint camelcase: ["error", {allow: ["codecept_helper"]}] */
@@ -12,26 +11,8 @@ chai.use(require('chai-sorted'));
 chai.use(require('chai-bignumber')());
 
 const expect = chai.expect;
-const otherFields = [
-	'username',
-	'limit',
-	'sort',
-	'offset',
-	'senderIdOrRecipientId',
-	'minAmount',
-	'maxAmount',
-	'fromTimestamp',
-	'toTimestamp',
-];
 
 class ValidateHelper extends Helper {
-	async getSchemaDefinition(name) {
-		const { result, error } = await from(liskCommons.schema());
-
-		expect(error).to.be.null;
-		return result.definitions[name];
-	}
-
 	async haveAccount(params) {
 		const { result, error } = await from(liskUtil.call().getAccounts(params));
 		const {
@@ -39,7 +20,10 @@ class ValidateHelper extends Helper {
 		} = result;
 
 		expect(error).to.be.null;
-		this.expectResponseToBeValid(result, 'AccountsResponse');
+		this.helpers.ResponseValidator.expectResponseToBeValid(
+			result,
+			'AccountsResponse'
+		);
 		return account;
 	}
 
@@ -76,7 +60,10 @@ class ValidateHelper extends Helper {
 		const account = await this.haveAccount({ address: params.address });
 
 		if (account && account.delegate) {
-			this.expectResponseToBeValid(account.delegate, 'Delegate');
+			this.helpers.ResponseValidator.expectResponseToBeValid(
+				account.delegate,
+				'Delegate'
+			);
 		} else {
 			if (account && account.secondPublicKey) {
 				delete params.secondPassphrase;
@@ -91,155 +78,6 @@ class ValidateHelper extends Helper {
 
 		await liskUtil.registerMultisignature(keepers, params);
 		return account;
-	}
-
-	async expectResponseToBeValid(response, definition) {
-		const schemaDefinition = await this.getSchemaDefinition(definition);
-
-		return expect(response).to.be.jsonSchema(schemaDefinition);
-	}
-
-	expectResponseToBeSortedBy(data, field, order) {
-		const result = data.map(item => item[field]);
-		if (result.length > 0 && !Number.isNaN(result[0])) {
-			return expect(result).to.deep.equal(sortBy(result, order));
-		}
-
-		if (order.toLowerCase() === 'asc') {
-			return expect(result).to.be.ascending;
-		}
-		return expect(result).to.be.descending;
-	}
-
-	expectResultToMatchParams(response, params) {
-		Object.entries(params).forEach(item => {
-			const [k, v] = item;
-			if (otherFields.includes(k)) {
-				this.handleOtherParams(response, k, v);
-			} else {
-				const data = flattern(response.data[0]);
-				expect(data[k].toString()).to.deep.equal(v);
-			}
-		});
-	}
-
-	expectBlockResultToMatchParams(response, params) {
-		const [[k, v]] = Object.entries(params);
-		if (['limit', 'sort', 'offset', 'blockId'].includes(k)) {
-			this.handleOtherParams(response, k, v);
-		} else {
-			const data = flattern(response.data[0]);
-			expect(data[k].toString()).to.deep.equal(v);
-		}
-	}
-
-	expectDelegatesToMatchParams(response, params) {
-		const [[k, v]] = Object.entries(params);
-		if (['limit', 'sort', 'offset', 'search'].includes(k)) {
-			this.handleOtherParams(response, k, v);
-		} else {
-			const data = flattern(response.data[0]);
-			expect(data[k].toString()).to.deep.equal(v);
-		}
-	}
-
-	expectMultisigAccountToHaveContracts(account, contracts) {
-		const addresses = account.data[0].members.map(m => m.address);
-		expect(contracts.every(c => addresses.includes(c.address))).to.deep.equal(
-			true
-		);
-	}
-
-	expectVotesResultToMatchParams(response, params) {
-		Object.entries(params).forEach(item => {
-			const [k, v] = item;
-			if (['sort'].includes(k)) {
-				const [field, order] = v.split(':');
-				this.expectResponseToBeSortedBy(response.data.votes, field, order);
-			} else if (k !== 'limit') {
-				expect(response.data[k].toString()).to.deep.equal(v);
-			}
-		});
-	}
-
-	expectVotersResultToMatchParams(response, params) {
-		Object.entries(params).forEach(item => {
-			const [k, v] = item;
-			if (['sort'].includes(k)) {
-				const [field, order] = v.split(':');
-				this.expectResponseToBeSortedBy(response.data.voters, field, order);
-			} else if (k !== 'limit') {
-				expect(response.data[k].toString()).to.deep.equal(v);
-			}
-		});
-	}
-
-	expectDefaultCount(response) {
-		expect(response.data.length).to.be.at.most(10);
-	}
-
-	handleOtherParams(response, key, value) {
-		switch (key) {
-			case 'username': {
-				expect(response.data[0].delegate.username).to.deep.equal(value);
-				break;
-			}
-			case 'limit':
-				expect(response.data).to.have.lengthOf(value);
-				break;
-
-			case 'offset':
-				expect(response.meta.offset).to.deep.equal(parseInt(value));
-				break;
-
-			case 'sort': {
-				const [field, order] = value.split(':');
-				this.expectResponseToBeSortedBy(response.data, field, order);
-				break;
-			}
-
-			case 'blockId':
-				expect(response.data[0].id).to.deep.equal(value);
-				break;
-
-			case 'search': {
-				response.data.forEach(element => {
-					const pattern = new RegExp(value);
-					expect(element.username).to.match(pattern);
-				});
-				break;
-			}
-			case 'senderIdOrRecipientId': {
-				const result = [
-					response.data[0].senderId,
-					response.data[0].recipientId,
-				];
-				expect(result).to.include(value);
-				break;
-			}
-			case 'minAmount':
-			case 'maxAmount':
-				response.data.forEach(t => {
-					if (key === 'minAmount') {
-						expect(t.amount).to.be.bignumber.at.least(value);
-					} else {
-						expect(t.amount).to.be.bignumber.at.most(value);
-					}
-				});
-				break;
-			case 'fromTimestamp':
-			case 'toTimestamp':
-				response.data.forEach(t => {
-					if (key === 'fromTimestamp') {
-						expect(t.timestamp).to.be.bignumber.at.least(value);
-					} else {
-						expect(t.timestamp).to.be.bignumber.at.most(value);
-					}
-				});
-				break;
-			default:
-				break;
-		}
 	}
 }
